Move JIRA REST API to it's own namespace
Maintain a better structuring by moving the API related classes into
it's own sub-namespace. Helps future refactoring.
