+++ features/bootstrap/Fixture/Tempo/TestHttpClient.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Fixture\Tempo;
+
+class TestHttpClient
+{
+
+}
