+++ src/Technodelight/Jira/Console/Command/SearchCommand.php
@@ -2,14 +2,12 @@
 
 namespace Technodelight\Jira\Console\Command;
 
-use GuzzleHttp\Exception\ClientException;
 use Symfony\Component\Console\Helper\QuestionHelper;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Question\Question;
-use Technodelight\Jira\Api\Api;
 use Technodelight\Jira\Console\Argument\NameNormalizer;
 
 class SearchCommand extends AbstractCommand
