+++ src/Technodelight/Jira/Console/Command/SelfUpdateCommand.php
@@ -2,7 +2,6 @@
 
 namespace Technodelight\Jira\Console\Command;
 
-use GuzzleHttp\Client;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
